requirements.txt
motor==3.3.1
pymongo==4.5.0
python-decouple==3.8
pytest==7.4.2
flask_jwt_extended=4.5.2
passlib=1.7.4
redis=5.0.0